requirement.txt
aiodns==2.0.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
brotlipy==0.7.0
cchardet==2.1.6
cffi==1.14.2
chardet==3.0.4
idna==2.10
multidict==4.7.6
pycares==3.1.1
pycparser==2.20
pycryptodome==3.9.8
steamid==1.0.2
vdf==3.3
yarl==1.5.1
